Task 0

+++ 0x00-python_variable_annotations/0-add.py
@@ -4,7 +4,7 @@
 """
 
 
-def add(a: int, b: int) -> float:
+def add(a: float, b: float) -> float:
     """
     a type-annotated function add that takes a float a and a float
     b as arguments and returns their sum as a float.
